# platform/krakend/openapi.yaml
openapi: 3.1.0
info:
  title: Fluxo de Caixa API
  version: 1.0.0
  description: Public gateway surface (/v1) — doc 02/05. KrakenD validates request bodies via JSON Schema derived from this spec.
servers:
  - url: https://krakend.gateway.svc.cluster.local:8080
    description: KrakenD gateway (cluster)
paths:
  /health:
    get:
      operationId: getHealth
      summary: Public gateway health (no JWT)
      responses:
        "200":
          description: OK
  /v1/lancamentos:
    get:
      operationId: listLancamentos
      summary: List lancamentos with filters and cursor (svc-lancamentos)
      security:
        - bearerAuth: []
      parameters:
        - name: data_inicio
          in: query
          schema:
            type: string
            format: date
        - name: data_fim
          in: query
          schema:
            type: string
            format: date
        - name: tipo
          in: query
          schema:
            type: string
            enum: [CREDITO, DEBITO]
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: Paginated list
        "401":
          description: Unauthorized
        "429":
          description: Rate limit exceeded
    post:
      operationId: createLancamento
      summary: Register lancamento (svc-lancamentos)
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LancamentoCreateRequest"
      responses:
        "201":
          description: Accepted
        "200":
          description: Idempotent replay
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "409":
          description: Idempotency conflict
        "429":
          description: Rate limit exceeded
  /v1/consolidado/{data}:
    get:
      operationId: getConsolidadoDiario
      summary: Daily consolidated balance (svc-consulta)
      security:
        - bearerAuth: []
      parameters:
        - name: data
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Consolidado for date
        "401":
          description: Unauthorized
        "429":
          description: Rate limit exceeded
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LancamentoCreateRequest:
      type: object
      additionalProperties: false
      required:
        - valor
        - tipo
      properties:
        valor:
          type: number
          exclusiveMinimum: 0
          maximum: 9999999.99
          multipleOf: 0.01
        tipo:
          type: string
          enum: [CREDITO, DEBITO]
        data_competencia:
          type: string
          format: date
        descricao:
          type: string
          maxLength: 500
        categoria_id:
          type: string
          format: uuid